# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "rustdesk-api-server",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/ab": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all address book and tags data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address book and tags"
                ],
                "summary": "Get all address book and tags data",
                "responses": {
                    "200": {
                        "description": "Response data is a serialized json string",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update all address book and tags data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address book and tags"
                ],
                "summary": "Update all address book and tags data",
                "parameters": [
                    {
                        "description": "Update data is a serialized json string",
                        "name": "AddressBookUpdateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddressBookController.AddressBookUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/ab/get": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all address book and tags data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Address book and tags"
                ],
                "summary": "Get all address book and tags data",
                "responses": {
                    "200": {
                        "description": "Response data is a serialized json string",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/activate/{token}": {
            "get": {
                "description": "User account activation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User account activation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activation token user received by email",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/audit": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "I don't know how to use it, and the request was not caught.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Audit logs",
                "responses": {
                    "200": {
                        "description": "Always return \"正常",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "I don't know how to use it, and the request was not caught.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Audit logs",
                "responses": {
                    "200": {
                        "description": "Always return \"正常",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/currentUser": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get current login user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "CurrentUser Info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserController.CurrentUserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/heartbeat": {
            "post": {
                "description": "Heartbeat with clients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Heartbeat with clients",
                "responses": {
                    "200": {
                        "description": "Always return unix timestamp of current time",
                        "schema": {
                            "$ref": "#/definitions/SystemController.HeartBeatResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "User login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login data",
                        "name": "LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserController.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserController.LoginResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Controllers.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/reg": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Username must be email, password min length is 8, name is required",
                        "name": "RegRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserController.RegRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        },
        "/resend/{username}": {
            "get": {
                "description": "User activation token resend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User activation token resend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username which you registered",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Controllers.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "AddressBookController.AddressBookUpdateRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "Controllers.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "Controllers.ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "SystemController.HeartBeatResponse": {
            "type": "object",
            "properties": {
                "modified_at": {
                    "type": "integer"
                }
            }
        },
        "UserController.CurrentUserResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "UserController.LoginRequest": {
            "type": "object",
            "properties": {
                "autoLogin": {
                    "type": "boolean"
                },
                "deviceInfo": {
                    "$ref": "#/definitions/UserController.LoginRequestDeviceInfo"
                },
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "UserController.LoginRequestDeviceInfo": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "UserController.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "UserController.RegRequest": {
            "type": "object",
            "required": [
                "name",
                "password",
                "username"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Retrieving access token by login api",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}